@@ -20,9 +20,8 @@
     "@types/jest": "24.9.1",
     "@types/markdown-it": "12.2.3",
     "@types/node": "16.18.38",
-    "@types/react": "17.0.2",
-    "@types/react-dom": "17.0.2",
-    "@types/react-redux": "7.1.22",
+    "@types/react": "18.2.18",
+    "@types/react-dom": "18.2.7",
     "@uiw/codemirror-extensions-langs": "4.12.4",
     "@uiw/codemirror-themes": "4.12.4",
     "@uiw/react-codemirror": "4.12.4",
@@ -31,7 +30,7 @@
     "@visx/hierarchy": "2.10.0",
     "@visx/responsive": "2.10.0",
     "@visx/scale": "2.2.2",
-    "antd": "4.20.4",
+    "antd": "4.24.5",
     "d3-dag": "0.9.1",
     "date-fns": "2.28.0",
     "fast-xml-parser": "4.2.4",
@@ -46,11 +45,11 @@
     "postman-collection": "4.1.4",
     "protobufjs": "7.2.4",
     "query-string": "7.1.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-flow-renderer": "10.3.7",
     "react-joyride": "2.5.3",
-    "react-redux": "7.2.6",
+    "react-redux": "8.1.2",
     "react-router-dom": "6.2.1",
     "react-scripts": "5.0.1",
     "react-spaces": "0.3.8",
