@@ -1,4 +1,4 @@
-import {createContext, useContext, useMemo, useState} from 'react';
+import {PropsWithChildren, createContext, useContext, useMemo, useState} from 'react';
 import {noop} from 'lodash';
 import * as S from './ContactUs.styled';
 import ContactUsModal from './ContactUsModal';
@@ -14,7 +14,7 @@ export const Context = createContext<IContext>({
 
 export const useContactUsModal = () => useContext(Context);
 
-const ContactUs: React.FC = ({children}) => {
+const ContactUs: React.FC<PropsWithChildren<{}>> = ({children}) => {
   const [isOpen, setIsOpen] = useState(false);
   const value = useMemo<IContext>(() => ({onOpen: () => setIsOpen(true)}), []);
 
