@@ -1,7 +1,8 @@
 import {ReadOutlined} from '@ant-design/icons';
+import {PropsWithChildren} from 'react';
 import * as S from './DocsBanner.styled';
 
-const DocsBanner: React.FC = ({children}) => {
+const DocsBanner: React.FC<PropsWithChildren<{}>> = ({children}) => {
   return (
     <S.DocsBannerContainer>
       <ReadOutlined />
