@@ -1,7 +1,8 @@
+import {PropsWithChildren} from 'react';
 import * as Spaces from 'react-spaces';
 import * as S from './ResizablePanels.styled';
 
-const ResizablePanels: React.FC = ({children}) => {
+const ResizablePanels: React.FC<PropsWithChildren<{}>> = ({children}) => {
   return (
     <>
       <S.GlobalStyle />
