@@ -12,7 +12,7 @@ interface IProps {
 
 const ResourceCardActions = ({id, shouldEdit = true, onDelete, onEdit}: IProps) => {
   const onDeleteClick = useCallback(
-    ({domEvent}) => {
+    ({domEvent}: {domEvent: React.MouseEvent<HTMLElement> | React.KeyboardEvent<HTMLElement>}) => {
       domEvent?.stopPropagation();
       onDelete();
     },
@@ -20,7 +20,7 @@ const ResourceCardActions = ({id, shouldEdit = true, onDelete, onEdit}: IProps) 
   );
 
   const onEditClick = useCallback(
-    ({domEvent}) => {
+    ({domEvent}: {domEvent: React.MouseEvent<HTMLElement> | React.KeyboardEvent<HTMLElement>}) => {
       domEvent?.stopPropagation();
       onEdit();
     },
@@ -28,17 +28,17 @@ const ResourceCardActions = ({id, shouldEdit = true, onDelete, onEdit}: IProps) 
   );
 
   const menuItems = useMemo(() => {
-    const defaultItems = [{key: 'delete', label: <span data-cy="test-card-delete">Delete</span>, onClick: onDeleteClick}];
+    const defaultItems = [
+      {key: 'delete', label: <span data-cy="test-card-delete">Delete</span>, onClick: onDeleteClick},
+    ];
 
-    return shouldEdit ? [{key: 'edit', label: <span data-cy="test-card-edit">Edit</span>, onClick: onEditClick}, ...defaultItems] : defaultItems;
+    return shouldEdit
+      ? [{key: 'edit', label: <span data-cy="test-card-edit">Edit</span>, onClick: onEditClick}, ...defaultItems]
+      : defaultItems;
   }, [onDeleteClick, onEditClick, shouldEdit]);
 
   return (
-    <Dropdown
-      overlay={<Menu items={menuItems} />}
-      placement="bottomLeft"
-      trigger={['click']}
-    >
+    <Dropdown overlay={<Menu items={menuItems} />} placement="bottomLeft" trigger={['click']}>
       <span data-cy={`test-actions-button-${id}`} className="ant-dropdown-link" onClick={e => e.stopPropagation()}>
         <S.ActionButton />
       </span>
